requirements.txt
Flask==3.0.1
flask_navigation==0.2.0
gunicorn==23.0.0

requests==2.32.4
structlog==25.1.0

google-auth==2.40.3

ddtrace==3.11.1

firebase-admin==7.0.0

google-cloud-logging==3.12.1
wtforms==3.2.1